requirements.txt
Flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.4
hypothesis==6.119.3
